+++ docs/changelog.md
@@ -1,4 +1,16 @@
 # TBD
+### Features
+* Added `KurtosisContext.getEnclaveContext` function
+* Added docs for `KurtosisContext`
+
+### Changes
+* `KurtosisContext.getEnclaves` now simply returns a set of enclave IDs
+
+### Breaking Changes
+* `NetworkContext` has been renamed to `EnclaveContext`
+* `KurtosisContext` enclave manipulation returns & accepts `EnclaveContext` objects from `kurtosis-client`
+* `KurtosisContext.getEnclaves` returns a set of enclave IDs
+* The `EnclaveInfo` object has an additional field, `enclaveDataDirpathOnHostMachine`
 
 # 0.9.0
 ### Removals